Seminarios/7 - AE/experiments/autoencoder_nlp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autoencoder-nlp"
version = "0.1.0"
description = "Autoencoder for NLP feature compression using Reuters RCV1 dataset"
requires-python = "==3.13.*"
dependencies = [
    "h5py>=3.15.1",
    "jupyterlab>=4.5.0",
    "lightning>=2.5.6",
    "matplotlib>=3.10.7",
    "mlflow>=2.19.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "typer>=0.20.0",
]

[project.scripts]
autoencoder_nlp = "autoencoder_nlp.cli:app"

[tool.setuptools]
packages = ["autoencoder_nlp"]

[tool.uv]
package = true
